+++ test/scittlets/reagent/test_utils.cljs
@@ -7,7 +7,7 @@
 
 (def scittlets (js->clj js/scittlets_metadata {:keywordize-keys true}))
 
-(println :scittlest scittlets)
+;;(println :scittlest scittlets)
 
 (defn- html-escape [s]
   (-> (clojure.string/replace s "&" "&amp;")
@@ -16,7 +16,7 @@
       (clojure.string/replace "\"" "&quot;")
       (clojure.string/replace "'" "&#39;")))
 
-(defn- url-open [url]
+(defn- url-text-open [url]
   (let [window-name url
         w (js/window.open "about:blank" window-name)]
     (-> (js/fetch url)
@@ -29,10 +29,11 @@
                    (.close doc)))))))
 
 (defn file-open+ [label url]
-  [:button {:on-click #(url-open url)
+  [:button {:on-click #(url-text-open url)
             :style {:all "unset"
                     :cursor "pointer"
                     :font-size "10px"
+                    :padding-right "5px"
                     :color "blue"
                     :text-decoration "underline"}}
    (str "[" label "]")])
@@ -53,16 +54,15 @@
                           :font-weight "bold"
                           :margin-right "4px"}}
            (if open? "▾" "▸")]
-          "Dependencies" (when-not open? " ...")]
+          (count deps) " Dependencies" (when-not open? " ...")]
          (when open?
            [:div
             [:pre {:style {:border "1px solid #ccc"
                            :margin-top "4px"
                            :padding "8px"
                            :background "#f9f9f9"
-                           :position "relative"}}  ;; Add relative positioning for the container
+                           :position "relative"}}
              sdeps
-             ;; Copy icon (positioned at the top-right corner)
              [:button
               {:on-click (fn []
                            (copy-to-clipboard sdeps)
@@ -79,20 +79,29 @@
                        :transition "color 0.3s ease"}}
               (if copied? "✔️" "📋")]]])]))))
 
-(defn API+ [var-fn]
+(defn API+ [var-fn home see]
   (let [{:keys [arglists doc file]
          nm :name nmspace :ns} (meta var-fn)
         name-full (str nmspace "/" nm)
         args (->> (first arglists)
                   (str/join " "))]
     [:div {:style {:font-family "Arial, sans-serif" :margin "1em"}}
-     [:h2 {:style {:color "#2c3e50"}} [:span name-full "  " [file-open+ "source" file]]]
+     [:h2 {:style {:color "#2c3e50"}} (into [:span name-full
+                                             "  " [:a {:href home :target "_blank"
+                                                       :style {:padding-right "5px"
+                                                               :font-size "15px"}}
+                                                   "🏠"]
+                                             [file-open+ "code" file]]
+                                            (map (fn [[label url]] (vector  :a {:href url :target "_blank"
+                                                                                :style {:padding-right "5px"
+                                                                                        :font-size "10px"}}
+                                                                            (str "[" (name label) "]"))) see))]
      [:p [:code (str "Usage: " "[" nm " " args "]") ]]
      [:pre {:style {:background "#f4f4f4" :padding "1em" :border-radius "5px"}}
       doc]]))
 
 (defn info+ [namespace-kw var-fn]
-  (let [{:keys [deps]} (get scittlets :scittlets.reagent.mermaid)]
+  (let [{:keys [deps home see]} (get scittlets :scittlets.reagent.mermaid)]
     [:<>
-     [API+ var-fn]
+     [API+ var-fn home see]
      [dependencies+ deps]]))