+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 ## Unreleased
+
+* Created Scittlets framework.
+* Added `scittlet.reagent.mermaid` scittlet.